requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
pyaspeller==2.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
